Update perf tool from iperf to qperf
- Update perf tool from iperf to qperf since the qperf
  provides richer metrics.
- Merge bandwidth_test.go into performance_test.go.
